Rename providers/config to pkg/credsfile
